Chat now has a container and is presentational
